% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicViewerSC.pckg.R
\name{addDf2seuratMetaData}
\alias{addDf2seuratMetaData}
\title{addDf2seuratMetaData}
\usage{
addDf2seuratMetaData(obj, dfAdd)
}
\arguments{
\item{obj}{Seurat object}
}
\value{
paramerer list
}
\description{
addDf2seuratMetaData
}
