% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.functions.R
\name{assignDbUsersAndPrivileges}
\alias{assignDbUsersAndPrivileges}
\title{assignDbUsersAndPrivileges}
\usage{
assignDbUsersAndPrivileges(
  accessFilePath = shinyDataPath,
  hostDbUrl = "10.27.241.82",
  appUserName = substr(paste0(project_id, "_aUser"), 1, 15),
  geneDefault = NULL,
  domains = c("shiny-bioinformatics.crick.ac.uk", "10.\%"),
  dbname = "prim.data.db",
  tables = c(coordTb = PCAdbTableName, exprTb = expDbTable, geneTb = geneTb),
  recreateProjectUser = TRUE,
  dbAdminUser = "boeings",
  dbAdminPwd = "db.pwd",
  dataMode = "MySQL"
)
}
\arguments{
\item{agree}{Do you agree dogs are the best pet? Defaults to TRUE.}
}
\description{
This function allows you to express your love for the superior furry animal.
}
\keyword{database}
\keyword{utils}
