% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.functions.R
\name{createAppPlotList}
\alias{createAppPlotList}
\title{createAppPlotList}
\usage{
createAppPlotList(input = "inputList", dfID = NULL, dfColOptions = NULL)
}
\arguments{
\item{input}{An input list}

\item{dfID}{A project parameter data frame}

\item{dfColOptions}{A color options data frame}
}
\value{
gg plot list
}
\description{
createAppPlotList
}
