% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.functions.R
\name{createDfCoordLocal}
\alias{createDfCoordLocal}
\title{createDfExprSelReport}
\usage{
createDfCoordLocal(dfID, dfColOptions = NULL, input = NULL)
}
\arguments{
\item{dfID}{An ID data frame}

\item{dfColOptions}{A color options table}

\item{input}{An input object}
}
\value{
dfCoordSel
}
\description{
createDfExprSelReport
}
