% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicViewerSC.pckg.R
\name{createDfExpr}
\alias{createDfExpr}
\title{createDfExpr}
\usage{
createDfExpr(obj, assay = "RNA", geneSel = NULL)
}
\arguments{
\item{obj}{Seurat object}

\item{assay}{Seurat assay to use to extract expression data}

\item{geneSel}{Gene subset to use when preparing the expression assay}
}
\value{
paramerer list
}
\description{
Scan Parameters
}
\details{
TBD.
}
