% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.functions.R
\name{createDfExprSelReport}
\alias{createDfExprSelReport}
\title{createDfExprSelReport}
\usage{
createDfExprSelReport(dfID, input)
}
\arguments{
\item{dfID}{An ID data frame}

\item{input}{An input object}
}
\value{
dfExprSel
}
\description{
createDfExprSelReport
}
