% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.functions.R
\name{createDfTempLocal}
\alias{createDfTempLocal}
\title{createDfTempLocal}
\usage{
createDfTempLocal(dfMeta, dfExpr, input, dfColorTable = NULL)
}
\arguments{
\item{dfMeta}{A Metadata table}

\item{dfExpr}{A color table}
}
\description{
createDfTempLocal
}
