% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.functions.R
\name{createInputListFromLink}
\alias{createInputListFromLink}
\title{createInputListFromLink}
\usage{
createInputListFromLink(figureLink = NULL)
}
\arguments{
\item{figureLink}{A figure link from the app}
}
\value{
input
}
\description{
createInputListFromLink
}
