% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.functions.R
\name{createPowerpointPresentation}
\alias{createPowerpointPresentation}
\title{createPowerpointPresentation}
\usage{
createPowerpointPresentation(
  slideList,
  outputfile = "singleCell.powerpoint.presentation.pptx"
)
}
\arguments{
\item{slideList}{A slide list}

\item{outputfile}{Specifies a powerpoint output file. Default is singleCell.powerpoint.presentation.pptx}
}
\description{
createPowerpointPresentation
}
