% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.functions.R
\name{dfColorTableReport}
\alias{dfColorTableReport}
\title{dfColorTableReport}
\usage{
dfColorTableReport(dfMeta, input, dfColOptions = NULL)
}
\arguments{
\item{dfMeta}{A meta data table}

\item{input}{An input object}

\item{dfColOptions}{A color options table}
}
\value{
dfDL
}
\description{
dfColorTableReport
}
