% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.functions.R
\name{plot_prep_report}
\alias{plot_prep_report}
\title{plot_prep_report}
\usage{
plot_prep_report(
  df,
  plot_name,
  colorBy = "lg10Expr",
  dotsize = 1,
  lowColor = "grey",
  dotcolor = "darkblue",
  x_axis = "UMAP_1",
  y_axis = "UMAP_2",
  background = "grey",
  maxX = NULL,
  minX = NULL,
  maxY = NULL,
  minY = NULL,
  geneSel = NULL,
  maxExpr = NULL,
  showPlotLegend = FALSE
)
}
\description{
plot_prep_report
}
