% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicViewerSC.pckg.R
\name{scanObjParams}
\alias{scanObjParams}
\title{Scan Parameters}
\usage{
scanObjParams(obj, NmaxSplit = 25, NcatColorMax = 25)
}
\arguments{
\item{obj}{Seurat object}

\item{NmaxSplit}{Maximum number of group members for a column/category to be listed in the splitBy column}

\item{NcatColorMax}{Value above which a numeric column is displayed as color gradient}
}
\value{
paramerer list
}
\description{
TBD.
}
