% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicViewerSC.pckg.R
\name{seuratObjectToLocalViewer}
\alias{seuratObjectToLocalViewer}
\title{seuratObjectToLocalViewer}
\usage{
seuratObjectToLocalViewer(
  params = NULL,
  project_id = "testApp",
  projectPath = "./",
  OsC = NULL,
  dataMode = "SQLite",
  geneDefault = NULL,
  dfExpr = NULL
)
}
\arguments{
\item{project_id}{Project id}

\item{OsC}{Seurat object}
}
\description{
seuratObjectToLocalViewer
}
