% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicViewerSC.pckg.R
\name{seuratObjectToViewer}
\alias{seuratObjectToViewer}
\title{seuratObjectToLocalViewer}
\usage{
seuratObjectToViewer(
  params = NULL,
  project_id = "testApp",
  projectPath = "./",
  OsC = NULL,
  dataMode = "MySQL",
  host = "dbHostURL",
  dbname = "dbname_db",
  db.pwd = "dbAdminPassword",
  db.user = "boeings",
  appDomains = c("shiny-bioinformatics.crick.ac.uk", "10.\%"),
  geneDefault = NULL,
  dfExpr = NULL
)
}
\arguments{
\item{project_id}{Project id}

\item{OsC}{Seurat object}
}
\description{
seuratObjectToLocalViewer
}
