% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.functions.R
\name{upload.datatable.to.database}
\alias{upload.datatable.to.database}
\title{upload.datatable.to.database}
\usage{
upload.datatable.to.database(
  host = NULL,
  user = NULL,
  password = NULL,
  prim.data.db = "project.database",
  dbTableName = "rnaseqdbTableName",
  df.data = "df.data.to.upload",
 
    db.col.parameter.list = list(`VARCHAR(255) CHARACTER SET latin1 COLLATE latin1_swedish_ci`
    = c("gene_description"), `VARCHAR(50) CHARACTER SET latin1 COLLATE latin1_swedish_ci`
    = c("ENSG", "ENSMUSG", "hgnc_symbol", "mgi_symbol", "uniprot", "entrezgene",
    "display_ptm", "^sequence_window", "p_site_env", "for_GSEA_gene_chip",
    "associated_gene_name", "gene_type"),
    `VARCHAR(1) CHARACTER SET latin1 COLLATE latin1_swedish_ci` = c("ppos", "amino_acid",
    "charge", "known_site"), `BIGINT(8) NULL DEFAULT NULL` = c("row_names"),
    `INT(8) NULL DEFAULT NULL` = c("row_id", 
     "cluster_order", "cluster_id",
    "count_cut_off", "^position$", "raw_counts"), `DECIMAL(6,3) NULL DEFAULT NULL` =
    c("norm_counts", "NES", "logFC", "lg2_avg", "intensity", "^int", "iBAQ",
    "^localization_prob$"), `DECIMAL(6,5) NULL DEFAULT NULL` = c("padj", "pvalue",
    "^pep$")),
  increment = 5000,
  new.table = FALSE,
  first.row.name.index = 1,
  startOnlyWithConnectionCount1 = FALSE,
  cols2Index = NULL,
  indexName = NULL,
  mode = "MySQL",
  addRowNamesColumn = TRUE
)
}
\arguments{
\item{host}{URL or IP address of the database server. NULL if mode is SQLite}

\item{user}{database user name. Needs privileges for INSERT, DELETE, DROP and SELECT}

\item{password}{database password}

\item{prim.data.db}{primary database name}

\item{dbTableName}{Name of the database table to upload}

\item{df.data}{data.frame to upload to the database}

\item{db.col.parameter.list}{This list specifies the category for a database column. This ideally is filled using the function inferDbColumns(dbTableName)}

\item{increment}{= 5000,}

\item{new.table}{= FALSE,}

\item{first.row.name.index}{= 1,}

\item{startOnlyWithConnectionCount1}{= FALSE,}

\item{cols2Index}{= NULL,}

\item{mode}{= "MySQL"}
}
\description{
mode options: SQLite, MySQL
}
\keyword{Database}
\keyword{MySQL}
\keyword{SQLite}
\keyword{Upload}
