% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicViewerSC.pckg.R
\name{writeAppParameterFiles}
\alias{writeAppParameterFiles}
\title{writeAppParameterFiles}
\usage{
writeAppParameterFiles(
  project_id = "testApp",
  projectPath = "./",
  params,
  menuParametersFN = "menuParameters.txt",
  colorParametersFN = "colorParameters.txt"
)
}
\arguments{
\item{project_id}{Project id}

\item{projectPath}{Path to project}

\item{params}{biologic parameterList}
}
\description{
writeAppParameterFiles
}
