% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.functions.R
\name{uploadDbTableInfile}
\alias{uploadDbTableInfile}
\title{uploadDbTablInfile}
\usage{
uploadDbTableInfile(
  host = NULL,
  user = NULL,
  password = NULL,
  prim.data.db = "project.database",
  dbTableName = "rnaseqdbTableName",
  df.data = "df.data.to.upload",
  db.col.parameter.list = list(`VARCHAR(255) CHARACTER SET latin1 COLLATE
    latin1_swedish_ci` = c("gene_description"), `VARCHAR(50) CHARACTER SET latin1 COLLATE
    latin1_swedish_ci` = c("ENSG", "ENSMUSG", "hgnc_symbol", "mgi_symbol", "uniprot",
    "entrezgene", "display_ptm", "^sequence_window", "p_site_env", "for_GSEA_gene_chip",
    "associated_gene_name", "gene_type"), `VARCHAR(1) CHARACTER SET latin1 COLLATE
    latin1_swedish_ci` = c("ppos", "amino_acid", "charge", "known_site"), `BIGINT(8) NULL
    DEFAULT NULL` = c("row_names"), `INT(8) NULL DEFAULT NULL` = c("row_id",     
    "cluster_order", "cluster_id", "count_cut_off", "^position$", "raw_counts"),
    `DECIMAL(6,3) NULL DEFAULT NULL` = c("norm_counts", "NES", "logFC", "lg2_avg",
    "intensity", "^int", "iBAQ", "^localization_prob$"), `DECIMAL(6,5) NULL DEFAULT NULL`
    = c("padj", "pvalue", "^pep$")),
  new.table = TRUE,
  cols2Index = NULL,
  indexName = NULL,
  mode = "MySQL",
  tempFileName = "temp.upload.csv"
)
}
\arguments{
\item{agree}{Do you agree dogs are the best pet? Defaults to TRUE.}
}
\description{
This function allows you to express your love for the superior furry animal.
}
\keyword{dogs}
